change chaining as default choice

+++ src/sylvan_obj.cpp
@@ -1084,7 +1084,7 @@ void Sylvan::printLevelToVar()
 {
     printf("level_to_var: ");
     for (size_t i = 0; i < sylvan_levelscount(); ++i){
-        printf("%u ", mtbdd_level_to_var(i));
+        printf("%zu ", (size_t)mtbdd_level_to_var(i));
     }
     printf("\n");
 }
@@ -1093,7 +1093,7 @@ void Sylvan::printVarToLevel()
 {
     printf("var_to_level: ");
     for (size_t i = 0; i < sylvan_levelscount(); ++i){
-        printf("%llu ", mtbdd_var_to_level(i));
+        printf("%zu ", (size_t)mtbdd_var_to_level(i));
     }
     printf("\n");
 }
